Move full cal load to cal model to remove qt dependencies from utils
get_setting (also imported in output_settings) imports qt
